Revert "always recreate symlink so buck2 always rebuilds chromium"
This reverts commit 07e2581454ba102262e43335a5ee1fcdb9d5a665.

+++ replay_build_scripts/common.mjs
@@ -129,7 +129,9 @@ export function updateBackendRepo() {
   // create a symlink to chromium in the backend checkout
   const chromiumRepoPath = process.cwd();
   const chromiumPathInBackend = path.join(backend, "chromium");
-  fs.unlinkSync(chromiumPathInBackend);
+  if (fs.existsSync(chromiumPathInBackend)) {
+    fs.unlinkSync(chromiumPathInBackend);
+  }
   fs.symlinkSync(chromiumRepoPath, chromiumPathInBackend);
 }
